requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
pynvml>=11.0.0
PyYAML>=5.4.1
tqdm>=4.62.0
pillow>=8.3.0
scikit-learn>=0.24.0
wandb>=0.12.0
pytest>=6.2.5
black>=21.7b0
isort>=5.9.3
flake8>=3.9.2
mypy>=0.910
